feat: #34번 티켓, 기존 부서 api에 employeeCount추가

+++ src/main/java/com/sprint/project1/hrbank/mapper/department/DepartmentMapper.java
@@ -15,7 +15,7 @@ import java.time.ZoneId;
 public interface DepartmentMapper {
 
     @Mapping(target = "establishedDate", source = "department.establishedDate", qualifiedByName = "changeLocalDateInSeoul")
-    DepartmentResponse toResponse(Department department, int employeeCount);
+    DepartmentResponse toResponse(Department department, long employeeCount);
 
 
     @Named("changeLocalDateInSeoul")
